+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@gesslar/aunty",
-  "version": "0.16.0",
+  "version": "0.17.0",
   "displayName": "Aunty Rose",
   "description": "Make gorgeous themes that speak as boldly as you do.",
   "publisher": "gesslar",
@@ -41,7 +41,7 @@
     "yaml"
   ],
   "dependencies": {
-    "ansi-colors": "^4.1.3",
+    "@gesslar/colours": "^0.0.1",
     "chokidar": "^4.0.3",
     "color-support": "^1.1.3",
     "commander": "^14.0.0",
